+++ demos/components-slider-gallery.php
@@ -69,7 +69,10 @@
                 <label for="gallery-a-4"><span class="prev">Previous</span><span class="next">Next</span></label>
                 <label for="gallery-a-5"><span class="prev">Previous</span><span class="next">Next</span></label>
               </div>
-              <div class="autoplay"><label for="gallery-a-autoplay"><span class="play"></span><span class="pause"></span></label></div>
+              <div class="autoplay">
+                <span class="autoplay-progress"></span>
+                <label for="gallery-a-autoplay"><span class="play"></span><span class="pause"></span></label>
+              </div>
             </div>
             <div class="gallery-b">
               <input type="radio" name="gallery-b" id="gallery-b-1" checked="">
@@ -82,27 +85,27 @@
                 <ul class="inner">
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x919/97cddd/97cddd"></a>
-                      <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>1 Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
-                      <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>2 Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x819/97cddd/97cddd"></a>
-                      <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>3 Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
-                      <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>4 Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x419/eea7d4/eea7d4"></a>
-                      <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>5 Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                 </ul>
@@ -135,27 +138,27 @@
                 <ul class="inner">
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
-                      <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>1 Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
-                      <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>2 Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
-                      <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>3 Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
-                      <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>4 Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                   <li>
                     <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
-                      <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                      <div class="info"><h3>5 Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
                     </div>
                   </li>
                 </ul>
@@ -181,7 +184,7 @@
         </div>
         <div class="content-aside">
           <pre class="language-scss"><code>
-.gallery-a { @include slider-gallery(5 autoplay default); }
+.gallery-a { @include slider-gallery(5 autoplay progress-bar default); }
 .gallery-b { @include slider-gallery(5 default); }
 .gallery-c { @include slider-gallery(5 default); }