+++ changelog.md
@@ -3,6 +3,7 @@
 ### v2.5.0
 + Add responsive font-size `font-size`
 + Add pure css `accordion`
++ Add `progress-bar` to css sliders
 
 ### v2.4.0
 + Re-add mixin `slider-carousel`, `slider-gallery`